Added a crypt service
