+++ Src/ToDoList/ToDoList.BLL/Services/UserService.cs
@@ -13,10 +13,12 @@ namespace ToDoList.BLL.Services
     {
         private const string EncryptionKey = "dsadasdsadas43";
         private readonly IUnitOfWork database;
+        private readonly ICryptService cryptService;
 
         public UserService()
         {
             database = new EFUnitOfWork();
+            cryptService = new CryptService();
         }
 
         public UserService(IUnitOfWork repository)
@@ -36,7 +38,7 @@ namespace ToDoList.BLL.Services
                 return Errors.UserExists;
             }
 
-            var user = new User { Password = Encrypt(password), UserName = userName, FullName = fullName };
+            var user = new User { Password = cryptService.Encrypt(password), UserName = userName, FullName = fullName };
             ((UserRepository)database.Users).Create(user);
             database.Save();
             AppConfig.UserId = user.Id;
@@ -52,7 +54,7 @@ namespace ToDoList.BLL.Services
             {
                 return Errors.Authentification;
             }
-            var decryptedPassword = Decrypt(user.Password);
+            var decryptedPassword = cryptService.Decrypt(user.Password);
             if (!decryptedPassword.Equals(password))
             {
                 return Errors.Authentification;
@@ -66,55 +68,7 @@ namespace ToDoList.BLL.Services
             return id == null ? string.Empty : database.Users.Get((int)id).FullName;
         }
 
-        private static string Encrypt(string clearText)
-        {
-            var clearBytes = Encoding.Unicode.GetBytes(clearText);
-            using (var encryptor = Aes.Create())
-            {
-                var pdb = new Rfc2898DeriveBytes(EncryptionKey, new byte[] { 0x49, 0x76, 0x61, 0x6e, 0x20, 0x4d, 0x65, 0x64, 0x76, 0x65, 0x64, 0x65, 0x76 });
-
-                if (encryptor == null) return clearText;
-
-                encryptor.Key = pdb.GetBytes(32);
-                encryptor.IV = pdb.GetBytes(16);
-                using (var ms = new MemoryStream())
-                {
-                    using (var cs = new CryptoStream(ms, encryptor.CreateEncryptor(), CryptoStreamMode.Write))
-                    {
-                        cs.Write(clearBytes, 0, clearBytes.Length);
-                        cs.Close();
-                    }
-
-                    clearText = Convert.ToBase64String(ms.ToArray());
-                }
-            }
-            return clearText;
-        }
-
-        private static string Decrypt(string cipherText)
-        {
-            var cipherBytes = Convert.FromBase64String(cipherText);
-            using (var encryptor = Aes.Create())
-            {
-                var pdb = new Rfc2898DeriveBytes(EncryptionKey, new byte[] { 0x49, 0x76, 0x61, 0x6e, 0x20, 0x4d, 0x65, 0x64, 0x76, 0x65, 0x64, 0x65, 0x76 });
-
-                if (encryptor == null) return cipherText;
-
-                encryptor.Key = pdb.GetBytes(32);
-                encryptor.IV = pdb.GetBytes(16);
-                using (var ms = new MemoryStream())
-                {
-                    using (var cs = new CryptoStream(ms, encryptor.CreateDecryptor(), CryptoStreamMode.Write))
-                    {
-                        cs.Write(cipherBytes, 0, cipherBytes.Length);
-                        cs.Close();
-                    }
-
-                    cipherText = Encoding.Unicode.GetString(ms.ToArray());
-                }
-            }
-            return cipherText;
-        }
+        
 
     }
 }
