+++ Src/ToDoList/ToDoList.BLL/ToDoList.BLL.csproj
@@ -22,6 +22,7 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <LangVersion>8.0</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>pdbonly</DebugType>
@@ -30,6 +31,7 @@
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <LangVersion>8.0</LangVersion>
   </PropertyGroup>
   <PropertyGroup>
     <StartupObject />
@@ -53,10 +55,12 @@
     <Compile Include="DTO\EventDto.cs" />
     <Compile Include="DTO\TaskDto.cs" />
     <Compile Include="DTO\UserDto.cs" />
+    <Compile Include="Interfaces\ICryptService.cs" />
     <Compile Include="Interfaces\IEventService.cs" />
     <Compile Include="Interfaces\INotificationService.cs" />
     <Compile Include="Interfaces\ITaskService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\CryptService.cs" />
     <Compile Include="Services\EventService.cs" />
     <Compile Include="Services\NotificationService.cs" />
     <Compile Include="Interfaces\IUserService.cs" />
